+++ asf_search/ASFProduct.py
@@ -129,10 +129,14 @@ class ASFProduct:
             - `path`: the expected path in the CMR UMM json granule response as a list
             - `cast`: (optional): the optional type casting method
 
+<<<<<<< HEAD
     Defining `_base_properties` in subclasses allows for
     defining custom properties or overiding existing ones.
     See `S1Product.get_property_paths()` on how subclasses are expected to
     combine `ASFProduct._base_properties` with their own separately defined `_base_properties`
+=======
+    Defining `_properties_paths` in subclasses allows for defining custom properties or overiding existing ones.
+>>>>>>> master
     """
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
@@ -195,23 +199,31 @@ class ASFProduct:
         if session is None:
             session = self.session
 
+        urls = self.get_urls(fileType=fileType)
+
+        for url in urls:
+            base_filename = '.'.join(default_filename.split('.')[:-1])
+            extension = url.split('.')[-1]
+            download_url(
+                url=url,
+                path=path,
+                filename=f"{base_filename}.{extension}",
+                session=session
+            )
+
+    def get_urls(self, fileType = FileDownloadType.DEFAULT_FILE) -> list:
         urls = []
 
         if fileType == FileDownloadType.DEFAULT_FILE:
-            urls.append((default_filename, self.properties["url"]))
+            urls.append(self.properties['url'])
         elif fileType == FileDownloadType.ADDITIONAL_FILES:
-            urls.extend(self._get_additional_filenames_and_urls(default_filename))
+            urls.extend(self.properties.get('additionalUrls', []))
         elif fileType == FileDownloadType.ALL_FILES:
-            urls.append((default_filename, self.properties["url"]))
-            urls.extend(self._get_additional_filenames_and_urls(default_filename))
+            urls.append(self.properties['url'])
+            urls.extend(self.properties.get('additionalUrls', []))
         else:
-            raise ValueError(
-                'Invalid FileDownloadType provided, '
-                'the valid types are "DEFAULT_FILE", "ADDITIONAL_FILES", and "ALL_FILES"'
-            )
-
-        for filename, url in urls:
-            download_url(url=url, path=path, filename=filename, session=session)
+            raise ValueError("Invalid FileDownloadType provided, the valid types are 'DEFAULT_FILE', 'ADDITIONAL_FILES', and 'ALL_FILES'")
+        return urls
 
     def _get_additional_filenames_and_urls(
         self,
@@ -341,9 +353,11 @@ class ASFProduct:
 
         umm = item.get("umm")
 
+        # additionalAttributes = {attr['Name']: attr['Values'] for attr in umm['AdditionalAttributes']}
+
         properties = {
-            prop: self._read_umm_property(umm, umm_mapping)
-            for prop, umm_mapping in self.get_property_paths().items()
+            prop:  self._read_umm_property(umm, umm_mapping)
+            for prop, umm_mapping in self._base_properties.items()
         }
 
         if properties.get("url") is not None:
@@ -361,22 +375,6 @@ class ASFProduct:
             properties["platform"] = self.umm_get(umm, "Platforms", 0, "ShortName")
 
         return {"geometry": geometry, "properties": properties, "type": "Feature"}
-
-    # ASFProduct subclasses define extra/override param key + UMM pathing here
-    @staticmethod
-    def get_property_paths() -> Dict:
-        """
-        Returns _base_properties of class, subclasses such as `S1Product` (or user provided subclasses) can override this to
-        define which properties they want in their subclass's properties dict.
-
-        (See `S1Product.get_property_paths()` for example of combining _base_properties of multiple classes)
-        
-        Returns
-        ----------
-        :dict
-            {`PROPERTY_NAME`: {'path': [umm, path, to, value], 'cast (optional)': Callable_to_cast_value}, ...} # noqa F401
-        """
-        return ASFProduct._base_properties
 
     def get_sort_keys(self) -> Tuple[str, str]:
         """
@@ -484,7 +482,9 @@ class ASFProduct:
         if item is None:
             return None
         for key in args:
-            if isinstance(key, int):
+            if isinstance(key, str):
+                item = item.get(key)
+            elif isinstance(key, int):
                 item = item[key] if key < len(item) else None
             elif isinstance(key, tuple):
                 (a, b) = key
@@ -507,8 +507,6 @@ class ASFProduct:
                         break
                 if not found:
                     return None
-            else:
-                item = item.get(key)
             if item is None:
                 return None
         if item in [None, "NA", "N/A", ""]: