+++ asf_search/ASFStackableProduct.py
@@ -14,17 +14,15 @@ class ASFStackableProduct(ASFProduct):
     ASF ERS-2 Dataset Documentation Page: https://asf.alaska.edu/datasets/daac/ers-2/
     """
 
-    _base_properties = {}
-
     class BaselineCalcType(Enum):
         """
         Defines how asf-search will calculate perpendicular baseline for products of this subclass
         """
 
         PRE_CALCULATED = 0
-        """Has pre-calculated insarBaseline value that will be used for perpendicular calculations""" # noqa F401
+        """Has pre-calculated insarBaseline value that will be used for perpendicular calculations"""  # noqa F401
         CALCULATED = 1
-        """Uses position/velocity state vectors and ascending node time for perpendicular calculations""" # noqa F401
+        """Uses position/velocity state vectors and ascending node time for perpendicular calculations"""  # noqa F401
 
     baseline_type = BaselineCalcType.PRE_CALCULATED
     """Determines how asf-search will attempt to stack products of this type."""
@@ -63,13 +61,6 @@ class ASFStackableProduct(ASFProduct):
         stack_opts.insarStackId = self.properties["insarStackId"]
         return stack_opts
 
-    @staticmethod
-    def get_property_paths() -> Dict:
-        return {
-            **ASFProduct.get_property_paths(),
-            **ASFStackableProduct._base_properties,
-        }
-
     def is_valid_reference(self):
         # we don't stack at all if any of stack is missing insarBaseline,
         # unlike stacking S1 products(?)
@@ -84,3 +75,8 @@ class ASFStackableProduct(ASFProduct):
         Returns the product type to search for when building a baseline stack.
         """
         return None
+
+    def has_baseline(self) -> bool:
+        baseline = self.get_baseline_calc_properties()
+
+        return baseline is not None