+++ asf_search/Products/ARIAS1GUNWProduct.py
@@ -15,20 +15,11 @@ class ARIAS1GUNWProduct(S1Product):
     """
 
     _base_properties = {
-        "perpendicularBaseline": {
-            "path": [
-                "AdditionalAttributes",
-                ("Name", "PERPENDICULAR_BASELINE"),
-                "Values",
-                0,
-            ],
-            "cast": try_parse_float,
-        },
-        "orbit": {"path": ["OrbitCalculatedSpatialDomains"]},
-        "inputGranules": {"path": ["InputGranules"]},
-        "ariaVersion": {
-            "path": ["AdditionalAttributes", ("Name", "VERSION"), "Values", 0]
-        },
+        **S1Product._base_properties,
+        'perpendicularBaseline': {'path': ['AdditionalAttributes', ('Name', 'PERPENDICULAR_BASELINE'), 'Values', 0], 'cast': try_parse_float},
+        'orbit': {'path': ['OrbitCalculatedSpatialDomains']},
+        'inputGranules': {'path': ['InputGranules']},
+        'ariaVersion': {'path': ['AdditionalAttributes', ('Name', 'VERSION'), 'Values', 0]}
     }
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
@@ -48,10 +39,6 @@ class ARIAS1GUNWProduct(S1Product):
                 self.properties["fileID"] + "." + urls[0].split(".")[-1]
             )
             self.properties["additionalUrls"] = urls[1:]
-
-    @staticmethod
-    def get_property_paths() -> Dict:
-        return {**S1Product.get_property_paths(), **ARIAS1GUNWProduct._base_properties}
 
     def get_stack_opts(self, opts: ASFSearchOptions = None) -> ASFSearchOptions:
         """