+++ asf_search/Products/ERSProduct.py
@@ -13,34 +13,17 @@ class ERSProduct(ASFStackableProduct):
     """
 
     _base_properties = {
-        "frameNumber": {
-            "path": ["AdditionalAttributes", ("Name", "FRAME_NUMBER"), "Values", 0]
-        },
-        "bytes": {
-            "path": ["AdditionalAttributes", ("Name", "BYTES"), "Values", 0],
-            "cast": try_round_float,
-        },
-        "esaFrame": {
-            "path": ["AdditionalAttributes", ("Name", "CENTER_ESA_FRAME"), "Values", 0]
-        },
-        "md5sum": {"path": ["AdditionalAttributes", ("Name", "MD5SUM"), "Values", 0]},
-        "beamModeType": {
-            "path": ["AdditionalAttributes", ("Name", "BEAM_MODE_TYPE"), "Values", 0]
-        },
-        "insarStackId": {
-            "path": ["AdditionalAttributes", ("Name", "INSAR_STACK_ID"), "Values", 0]
-        },
+        **ASFStackableProduct._base_properties,
+        'frameNumber': {'path': ['AdditionalAttributes', ('Name', 'FRAME_NUMBER'), 'Values', 0]},
+        'bytes': {'path': ['AdditionalAttributes', ('Name', 'BYTES'), 'Values', 0], 'cast': try_round_float},
+        'esaFrame': {'path': ['AdditionalAttributes', ('Name', 'CENTER_ESA_FRAME'), 'Values', 0]},
+        'md5sum': {'path': ['AdditionalAttributes', ('Name', 'MD5SUM'), 'Values', 0]},
+        'beamModeType': {'path': ['AdditionalAttributes', ('Name', 'BEAM_MODE_TYPE'), 'Values', 0]},
+        'insarStackId': {'path': ['AdditionalAttributes', ('Name', 'INSAR_STACK_ID'), 'Values', 0]},
     }
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
         super().__init__(args, session)
-
-    @staticmethod
-    def get_property_paths() -> Dict:
-        return {
-            **ASFStackableProduct.get_property_paths(),
-            **ERSProduct._base_properties,
-        }
 
     @staticmethod
     def get_default_baseline_product_type() -> Union[str, None]: