+++ asf_search/Products/NISARProduct.py
@@ -8,8 +8,10 @@ class NISARProduct(ASFStackableProduct):
 
     ASF Dataset Documentation Page: https://asf.alaska.edu/nisar/
     """
-
-    _base_properties = {"pgeVersion": {"path": ["PGEVersionClass", "PGEVersion"]}}
+    _base_properties = {
+        **ASFStackableProduct._base_properties,
+        'pgeVersion': {'path': ['PGEVersionClass', 'PGEVersion']}
+    }
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
         super().__init__(args, session)
@@ -38,13 +40,6 @@ class NISARProduct(ASFStackableProduct):
         for building a stack from this product
         """
         return None
-
-    @staticmethod
-    def get_property_paths() -> Dict:
-        return {
-            **ASFStackableProduct.get_property_paths(),
-            **NISARProduct._base_properties,
-        }
 
     def get_sort_keys(self) -> Tuple[str, str]:
         keys = super().get_sort_keys()