+++ asf_search/Products/RADARSATProduct.py
@@ -1,6 +1,11 @@
 from typing import Dict, Union
+<<<<<<< HEAD
 from asf_search import ASFSession, ASFStackableProduct
 from asf_search.CMR.translate import try_parse_float
+=======
+from asf_search import ASFSearchOptions, ASFSession, ASFProduct, ASFStackableProduct
+from asf_search.CMR.translate import try_parse_float, try_parse_int
+>>>>>>> master
 from asf_search.constants import PRODUCT_TYPE
 
 
@@ -10,6 +15,7 @@ class RADARSATProduct(ASFStackableProduct):
     """
 
     _base_properties = {
+<<<<<<< HEAD
         "faradayRotation": {
             "path": ["AdditionalAttributes", ("Name", "FARADAY_ROTATION"), "Values", 0],
             "cast": try_parse_float,
@@ -21,12 +27,22 @@ class RADARSATProduct(ASFStackableProduct):
         "insarStackId": {
             "path": ["AdditionalAttributes", ("Name", "INSAR_STACK_ID"), "Values", 0]
         },
+=======
+        **ASFStackableProduct._base_properties,
+        'faradayRotation': {'path': ['AdditionalAttributes', ('Name', 'FARADAY_ROTATION'), 'Values', 0], 'cast': try_parse_float},
+        'md5sum': {'path': ['AdditionalAttributes', ('Name', 'MD5SUM'), 'Values', 0]},
+        'beamModeType': {'path': ['AdditionalAttributes', ('Name', 'BEAM_MODE_TYPE'), 'Values', 0]},
+        'insarStackId': {'path': ['AdditionalAttributes', ('Name', 'INSAR_STACK_ID'), 'Values', 0]},
+        'frameNumber': {'path': ['AdditionalAttributes', ('Name', 'FRAME_NUMBER'), 'Values', 0], 'cast': try_parse_int}, #Sentinel and ALOS product alt for frameNumber (ESA_FRAME)
+        'esaFrame': {'path': ['AdditionalAttributes', ('Name', 'CENTER_ESA_FRAME'), 'Values', 0], 'cast': try_parse_int},
+>>>>>>> master
     }
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
         super().__init__(args, session)
 
     @staticmethod
+<<<<<<< HEAD
     def get_property_paths() -> Dict:
         return {
             **ASFStackableProduct.get_property_paths(),
@@ -34,6 +50,8 @@ class RADARSATProduct(ASFStackableProduct):
         }
 
     @staticmethod
+=======
+>>>>>>> master
     def get_default_baseline_product_type() -> Union[str, None]:
         """
         Returns the product type to search for when building a baseline stack.