+++ asf_search/Products/SIRCProduct.py
@@ -8,6 +8,7 @@ class SIRCProduct(ASFProduct):
     """
 
     _base_properties = {
+<<<<<<< HEAD
         "groupID": {
             "path": ["AdditionalAttributes", ("Name", "GROUP_ID"), "Values", 0]
         },
@@ -16,11 +17,21 @@ class SIRCProduct(ASFProduct):
         "beamModeType": {
             "path": ["AdditionalAttributes", ("Name", "BEAM_MODE_TYPE"), "Values", 0]
         },
+=======
+        **ASFProduct._base_properties,
+        'groupID': {'path': [ 'AdditionalAttributes', ('Name', 'GROUP_ID'), 'Values', 0]},
+        'md5sum': {'path': [ 'AdditionalAttributes', ('Name', 'MD5SUM'), 'Values', 0]},
+        'pgeVersion': {'path': ['PGEVersionClass', 'PGEVersion'] },
+        'beamModeType': {'path': ['AdditionalAttributes', ('Name', 'BEAM_MODE_TYPE'), 'Values', 0]},
+>>>>>>> master
     }
 
     def __init__(self, args: Dict = {}, session: ASFSession = ASFSession()):
         super().__init__(args, session)
+<<<<<<< HEAD
 
     @staticmethod
     def get_property_paths() -> Dict:
         return {**ASFProduct.get_property_paths(), **SIRCProduct._base_properties}
+=======
+>>>>>>> master