+++ asf_search/export/jsonlite2.py
@@ -5,6 +5,7 @@ from types import GeneratorType
 from asf_search import ASF_LOGGER
 from .jsonlite import JSONLiteStreamArray
 
+<<<<<<< HEAD
 
 def results_to_jsonlite2(results):
     ASF_LOGGER.info("started translating results to jsonlite2 format")
@@ -12,6 +13,16 @@ def results_to_jsonlite2(results):
     if not inspect.isgeneratorfunction(results) and not isinstance(
         results, GeneratorType
     ):
+=======
+def results_to_jsonlite2(results):
+    ASF_LOGGER.info('started translating results to jsonlite2 format')
+    
+    if len(results) == 0:
+            yield from json.JSONEncoder(indent=2, sort_keys=True).iterencode({'results': []})
+            return
+        
+    if not inspect.isgeneratorfunction(results) and not isinstance(results, GeneratorType):
+>>>>>>> master
         results = [results]
 
     streamer = JSONLite2StreamArray(results)
@@ -64,6 +75,7 @@ class JSONLite2StreamArray(JSONLiteStreamArray):
             "pge": p["pgeVersion"],
         }
 
+<<<<<<< HEAD
         if "temporalBaseline" in p.keys() or "perpendicularBaseline" in p.keys():
             result["tb"] = p["temporalBaseline"]
             result["pb"] = p["perpendicularBaseline"]
@@ -71,6 +83,19 @@ class JSONLite2StreamArray(JSONLiteStreamArray):
         if p.get("burst") is not None:  # is a burst product
             result["s1b"] = p["burst"]
 
+=======
+        if 'temporalBaseline' in p.keys():
+            result['tb'] = p['temporalBaseline']
+        if 'perpendicularBaseline' in p.keys():
+            result['pb'] = p['perpendicularBaseline']
+
+        if p.get('burst') is not None: # is a burst product
+            result['s1b'] = p['burst']
+
+        if p.get('opera') is not None:
+            result['s1o'] = p['opera']
+        
+>>>>>>> master
         return result
 
     def getOutputType(self) -> str: