+++ asf_search/search/baseline_search.py
@@ -126,16 +126,8 @@ def _cast_to_subclass(product: ASFProduct, subclass: Type[ASFProduct]) -> ASFPro
     ```
     class MyCustomClass(ASFProduct):
         _base_properties = {
+        **ASFProduct._base_properties,
         'some_unique_property': {'path': ['AdditionalAttributes', 'UNIQUE_PROPERTY', ...]}
-        }
-
-        ...
-
-        @staticmethod
-        def get_property_paths() -> dict:
-        return {
-            **ASFProduct.get_property_paths(),
-            **MyCustomClass._base_properties
         }
 
     # subclass as constructor