+++ asf_search/search/search.py
@@ -1,3 +1,4 @@
+import time
 from typing import Union, Sequence, Tuple
 from copy import copy
 import datetime
@@ -173,10 +174,15 @@ def search(
     results = ASFSearchResults([])
 
     # The last page will be marked as complete if results sucessful
+    perf = time.time()
     for page in search_generator(opts=opts):
+        ASF_LOGGER.warning(f"Page Time Elapsed {time.time() - perf}")
         results.extend(page)
         results.searchComplete = page.searchComplete
         results.searchOptions = page.searchOptions
+        perf = time.time()
+
+    results.raise_if_incomplete()
 
     try:
         results.sort(key=lambda p: p.get_sort_keys(), reverse=True)